transfered some of the function of opsim_pipeline module to util_lsst.py
