+++ tests/test_LsstSciencePipeline/test_opsim_pipeline.py
@@ -3,10 +3,8 @@ import numpy as np
 from astropy.table import Table
 from astropy.cosmology import FlatLambdaCDM
 from slsim.lens import Lens
-from slsim.LsstSciencePipeline.opsim_pipeline import (
-    opsim_variable_lens_injection,
-    opsim_time_series_images_data,
-)
+from slsim.LsstSciencePipeline.opsim_pipeline import opsim_time_series_images_data
+from slsim.LsstSciencePipeline.util_lsst import opsim_variable_lens_injection
 from slsim.Sources.source import Source
 from slsim.Deflectors.deflector import Deflector
 import pytest
